arch/x86/cpu_info: Fix compilation with recent GCC versions

+++ src/arch/x86/lib/cpu_performance/cpu_counter.c
@@ -11,10 +11,10 @@
 #include <hal/cpu_info.h>
 
 uint64_t get_cpu_counter(void) {
-	uint64_t hi = 0, lo = 0;
+	uint32_t hi = 0, lo = 0;
 	asm volatile (	"rdtsc\n\t"
 					"movl %%eax, %0\n\t"
 					"movl %%edx, %1\n\t" :
 					"=r"(lo), "=r"(hi) :);
-	return (hi << 32) + lo;
+	return (((uint64_t) hi) << 32) + lo;
 }
